+++ crates/router/tests/payments2.rs
@@ -206,7 +206,9 @@ async fn payments_create_core() {
         frm_metadata: None,
         merchant_order_reference_id: None,
         order_tax_amount: None,
+        connector_mandate_id: None,
     };
+
     let expected_response =
         services::ApplicationResponse::JsonWithHeaders((expected_response, vec![]));
     let actual_response = Box::pin(payments::payments_core::<
@@ -471,6 +473,7 @@ async fn payments_create_core_adyen_no_redirect() {
             frm_metadata: None,
             merchant_order_reference_id: None,
             order_tax_amount: None,
+            connector_mandate_id: None,
         },
         vec![],
     ));